requirements.txt
kerykeion
requests
pandas
boto3
python-dotenv
streamlit
